backend/requirements.txt
fastapi
requests
uvicorn
python-dotenv
Flask
google-generativeai
flask-cors 
gspread
google-auth
python-jose[cryptography]
passlib[bcrypt]
requests-toolbelt 